requirements.txt
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain_core
langchain-groq>=0.2,<0.3
pydantic>=2,<3
pydantic-core>=2,<3
pydantic-settings>=2,<3
python-dotenv>=1.0.0
ipykernel>=6.29.0
pandas>=2.2.1
mysql-connector-python>=8.3.0
SQLAlchemy>=2.0.25
streamlit>=1.32.0
psycopg2-binary
